refactor: move row->Parquet materialization onto the DeltaHarness SPI
Delete the standalone RowParquet object; add schemaAt / writeRows /
writeRowsToTemp to the DeltaHarness trait (implemented in DeltaSparkHarness, with
the value coercion kept private and unchanged). Rewire the capture-side callers
(WorkloadContext, WriteSpecCapture) to the harness. Consolidates this Spark-only
capture logic behind the SPI and removes a file.
